requirements.txt
PySide6>=6.6
fastapi>=0.110
uvicorn>=0.29
pyyaml>=6.0
requests>=2.31
